feed/pyproject.toml
[project]
name = "octogen-showcase-feed"
version = "0.1.0"
description = "Showcase of Octogen Feed generation agent"
requires-python = ">=3.12"
dependencies = [
    "fastapi-cli>=0.0.7",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "octogen-api>=0.1.0a8",
    "octogen-sdk-langgraph>=0.1.8",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/octogen"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
mypy_path = "src"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.1",
]
